submitted

+++ ex01/inc/PhoneBook.hpp
@@ -6,7 +6,7 @@
 /*   By: myuen <myuen@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/11 16:53:29 by myuen             #+#    #+#             */
-/*   Updated: 2025/05/15 19:21:04 by myuen            ###   ########.fr       */
+/*   Updated: 2025/05/17 16:02:54 by myuen            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,9 +22,9 @@
 class PhoneBook
 {
 private:
-	static const int	MAX_CONTACTS = 8;
+	static const size_t	MAX_CONTACTS = 8;
 	Contact				contacts[MAX_CONTACTS];
-	int					count;
+	size_t				count;
 	
 	std::string truncate(const std::string& str) const;
 public:
@@ -33,9 +33,8 @@ public:
 
 	void	add_contact(Contact new_contact);
 	void	print_all () const;
-	void	print_a_contact (int index) const;
+	void	print_a_contact (size_t index) const;
 	int 	get_count(void) const;
-	
 };
 
 #endif
